+++ cognite/Cargo.toml
@@ -1,25 +1,29 @@
 [package]
+authors = [
+    "Einar Marstrander Omang <einar.omang@cognite.com>",
+    "Haakon Garseg Mørk <haakon.mork@cognite.com>",
+    "Niek Beckers <niek.beckers@cognite.com>",
+]
+edition = "2021"
 name = "cognite"
-version = "0.2.0"
-authors = ["Haakon Garseg Mørk <haakon.mork@cognite.com>", "Einar Marstrander Omang <einar.omang@cognite.com>"]
 publish = false
-edition = "2018"
+version = "0.2.1"
 
 [dependencies]
-reqwest = { version = "0.11.14", features = ["json", "stream"] }
-serde = { version = "1.0.152", features = ["derive"] }
-serde_json = "1.0.94"
-uuid = { version = "1.3.0", features = ["serde", "v4"] }
+async-trait = "0.1.66"
+bytes = "1.4.0"
+futures = "0.3.26"
+futures-locks = "0.7.1"
+futures-timer = "3.0.2"
 prost = "0.11.8"
 prost-types = "0.11.8"
-futures-locks = "0.7.1"
-async-trait = "0.1.66"
-futures = "0.3.26"
-thiserror = "1.0.39"
-bytes = "1.4.0"
+reqwest = { version = "0.11.14", features = ["json", "stream"] }
 reqwest-middleware = "0.2.0"
+serde = { version = "1.0.152", features = ["derive"] }
+serde_json = "1.0.94"
 task-local-extensions = "0.1.3"
-futures-timer = "3.0.2"
+thiserror = "1.0.39"
+uuid = { version = "1.3.0", features = ["serde", "v4"] }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
 tokio = { version = "1.26.0", features = ["macros", "rt-multi-thread"] }
