+++ cognite/tests/data_modeling_tests.rs
@@ -1,0 +1,5 @@
+pub mod data_modeling_tests {
+    pub mod instances_tests;
+}
+
+pub mod common;
